Moved folder and improved code
