+++ js/core/classes/coin.js
