+++ js/core/classes/particles.js
