+++ js/core/classes/sound.js
