+++ js/core/index.js
@@ -1,8 +1,8 @@
 
-import ParticleSystem from './source/particles.js';
-import Player from './source/player.js';
-import Enemy from './source/enemies.js';
-import Coin from './source/coin.js';
+import ParticleSystem from './classes/particles.js';
+import Player from './classes/player.js';
+import Enemy from './classes/enemies.js';
+import Coin from './classes/coin.js';
 
 export {
   ParticleSystem,
